fix_view2

+++ app/models/cart.rb
@@ -13,10 +13,12 @@ class Cart < ApplicationRecord
 
   def del_product(product_id)
     current_item = line_items.find_by(product_id: product_id)
-    if current_item.quantity > 1
-      current_item.quantity -= 1
-    else
-      current_item.destroy
+    unless current_item.nil?
+      if current_item.quantity > 1
+        current_item.quantity -= 1
+      else
+        current_item.destroy
+      end
     end
     current_item
   end
